Fixed some strange warning on Windows..

+++ src/SolvingManager.cpp
@@ -77,7 +77,7 @@ static void solve(ConcurrentQueue<EquationCoefficients>& equationCoefQueue, Cons
             continue;
         }
 
-        long long discriminant = std::pow(b, 2) - 4LL * a * c;
+        long long discriminant = static_cast<long long>(std::pow(b, 2)) - 4LL * a * c;
         if (discriminant < 0) {
             oss << "(no roots) ";
             findExtremum(oss, a, b, c);
